New input system full support with touch and pen inputs

+++ Assets/Scripts/CommandPattern/CommandHandler.cs
@@ -1,25 +1,40 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.UI;
 
 public class CommandHandler
 {
     public List<ICommand> commandList = new List<ICommand>();
     public List<Vector3Int> selectedLocations = new List<Vector3Int>();
+    public List<float> riskValues = new List<float>();
+    public int batteryLevel = 12;
+    public int batteryMax = 12;
     public float accumulatedRisk = 0;
     public int index;
 
-    public void AddCommand(ICommand command)
-    {       
-        {if (index < commandList.Count)
-            commandList.RemoveRange(index, commandList.Count - index);
-            selectedLocations.RemoveRange(index, selectedLocations.Count - index);
+    // For the risk graph
+    private Window_Graph windowGraph;
 
-        commandList.Add(command);
-        selectedLocations.Add(command.clickedLocation);
-        addRisk(command.clickedLocation);
-        command.Execute();
-        index++;
+    public void AddCommand(ICommand command)
+    {
+        if (batteryLevel > 0)
+        {
+            {if (index < commandList.Count)
+                commandList.RemoveRange(index, commandList.Count - index);
+                selectedLocations.RemoveRange(index, selectedLocations.Count - index);
+
+            commandList.Add(command);
+            selectedLocations.Add(command.clickedLocation);
+            addRisk(command.clickedLocation);
+            removeEnergy(1);
+            command.Execute();
+            index++;
+            }
+        }
+        else
+        {
+            Debug.Log("Battery is empty");
         }
     }
 
@@ -27,11 +42,14 @@ public class CommandHandler
     {
         if (commandList.Count == 0)
             return;
-        if (index > 0)
-        {   
+        if (index > 1)
+        {
+            // not able to undo a given command...
             commandList[index - 1].Undo();
+            Debug.Log($"Last selected: {selectedLocations[index-1]} and index is {index}. Deleting {selectedLocations[index - 1]}");
             selectedLocations.RemoveAt(index - 1);
             removeRisk(commandList[index - 1].clickedLocation); // not sure if index is 1 or 0
+            addEnergy(1);
             index--;
         }
         Debug.Log("Command removed");
@@ -39,14 +57,72 @@ public class CommandHandler
 
     void addRisk(Vector3Int tileLocalPos)
     {
-        accumulatedRisk += commandList[index].gridInfo.GetPositionProperty(tileLocalPos, "Risk", 1.0f);
-        // Debug.Log($"Accumulated risk is {accumulatedRisk}");
+        accumulatedRisk += commandList[index].gridInfo.GetPositionProperty(tileLocalPos, "Risk", 0.0f);
+        riskValues.Add(accumulatedRisk);
+        ShowRiskGraph(riskValues);
     }
 
     void removeRisk(Vector3Int tileLocalPos)
     {
-        accumulatedRisk -= commandList[index-1].gridInfo.GetPositionProperty(tileLocalPos, "Risk", 1.0f);
-        // Debug.Log($"Accumulated risk is {accumulatedRisk}");
+        accumulatedRisk -= commandList[index-1].gridInfo.GetPositionProperty(tileLocalPos, "Risk", 0.0f);
+        riskValues.RemoveAt(index-1);
+        ShowRiskGraph(riskValues);
+    }
+
+    public void ShowRiskGraph(List<float> riskValues)
+    {
+        windowGraph = Transform.FindObjectOfType<Window_Graph>();
+        windowGraph.ShowGraph(riskValues);
+    }
+    
+
+    void addEnergy(int energy)
+    {
+        batteryLevel += energy;
+        Debug.Log($"adding energy, current level: {batteryLevel}");
+        if(batteryLevel < 0.7*batteryMax)
+        {
+            Image batteryLevelText = GameObject.Find("Full").GetComponent<Image>();
+            batteryLevelText.sprite = Resources.Load<Sprite>("Batteries/Half");
+            if (batteryLevel < 0.3*batteryMax)
+            {
+                batteryLevelText.sprite = Resources.Load<Sprite>("Batteries/Low");
+                if (batteryLevel <= 0)
+                {
+                    batteryLevelText.sprite = Resources.Load<Sprite>("Batteries/Zero");
+                }
+            }
+        }
+        else
+        {
+            Image batteryLevelText = GameObject.Find("Full").GetComponent<Image>();
+            batteryLevelText.sprite = Resources.Load<Sprite>("Batteries/Full");
+        }
+
+    }
+    void removeEnergy(int energy)
+    {
+        batteryLevel -= energy;
+        if(batteryLevel < 0.7*batteryMax)
+        {
+            Image batteryLevelText = GameObject.Find("Full").GetComponent<Image>();
+            batteryLevelText.sprite = Resources.Load<Sprite>("Batteries/Half");
+            if (batteryLevel < 0.3*batteryMax)
+            {
+                batteryLevelText.sprite = Resources.Load<Sprite>("Batteries/Low");
+                if (batteryLevel <= 0)
+                {
+                    batteryLevelText.sprite = Resources.Load<Sprite>("Batteries/Zero");
+                }
+            }
+
+        }
+        else
+        {
+            Image batteryLevelText = GameObject.Find("Full").GetComponent<Image>();
+            batteryLevelText.sprite = Resources.Load<Sprite>("Batteries/Full");
+        }
+        Debug.Log($"removing energy, current level: {batteryLevel}");
     }
 }
